Refactor transaction schema exports
