+++ packages/schemas/src/index.ts
@@ -1,5 +1,1 @@
-export {
-	s2cTransactionSchemaRaw,
-	s2cTransactionSchema,
-	c2sTransactionSchema,
-} from './transaction';
+export * from './transaction';
